Split off creating the Controller from the Reconciler.
