+++ pkg/reconciler/eventtype/eventtype.go
@@ -28,7 +28,6 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
-	eventinginformers "github.com/knative/eventing/pkg/client/informers/externalversions/eventing/v1alpha1"
 	listers "github.com/knative/eventing/pkg/client/listers/eventing/v1alpha1"
 	"github.com/knative/eventing/pkg/logging"
 	"github.com/knative/eventing/pkg/reconciler"
@@ -39,12 +38,6 @@ import (
 )
 
 const (
-	// ReconcilerName is the name of the reconciler.
-	ReconcilerName = "EventTypes"
-	// controllerAgentName is the string used by this controller to identify
-	// itself when creating events.
-	controllerAgentName = "eventtype-controller"
-
 	// Name of the corev1.Events emitted from the reconciliation process.
 	eventTypeReadinessChanged   = "EventTypeReadinessChanged"
 	eventTypeReconcileFailed    = "EventTypeReconcileFailed"
@@ -64,37 +57,6 @@ var brokerGVK = v1alpha1.SchemeGroupVersion.WithKind("Broker")
 
 // Check that our Reconciler implements controller.Reconciler
 var _ controller.Reconciler = (*Reconciler)(nil)
-
-// NewController initializes the controller and is called by the generated code
-// Registers event handlers to enqueue events
-func NewController(
-	opt reconciler.Options,
-	eventTypeInformer eventinginformers.EventTypeInformer,
-	brokerInformer eventinginformers.BrokerInformer,
-) *controller.Impl {
-
-	r := &Reconciler{
-		Base:            reconciler.NewBase(opt, controllerAgentName),
-		eventTypeLister: eventTypeInformer.Lister(),
-		brokerLister:    brokerInformer.Lister(),
-	}
-	impl := controller.NewImpl(r, r.Logger, ReconcilerName)
-
-	r.Logger.Info("Setting up event handlers")
-	eventTypeInformer.Informer().AddEventHandler(controller.HandleAll(impl.Enqueue))
-
-	// Tracker is used to notify us that a EventType's Broker has changed so that
-	// we can reconcile.
-	r.tracker = tracker.New(impl.EnqueueKey, opt.GetTrackerLease())
-	brokerInformer.Informer().AddEventHandler(controller.HandleAll(
-		controller.EnsureTypeMeta(
-			r.tracker.OnChanged,
-			v1alpha1.SchemeGroupVersion.WithKind("Broker"),
-		),
-	))
-
-	return impl
-}
 
 // Reconcile compares the actual state with the desired, and attempts to
 // converge the two. It then updates the Status block of the EventType resource
