+++ pkg/reconciler/inmemorychannel/controller.go
@@ -1,0 +1,82 @@
+/*
+Copyright 2019 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package inmemorychannel
+
+import (
+	messaginginformers "github.com/knative/eventing/pkg/client/informers/externalversions/messaging/v1alpha1"
+	"github.com/knative/eventing/pkg/reconciler"
+	"github.com/knative/pkg/controller"
+	appsv1informers "k8s.io/client-go/informers/apps/v1"
+	corev1informers "k8s.io/client-go/informers/core/v1"
+	"k8s.io/client-go/tools/cache"
+)
+
+const (
+	// ReconcilerName is the name of the reconciler
+	ReconcilerName = "InMemoryChannels"
+
+	// controllerAgentName is the string used by this controller to identify
+	// itself when creating events.
+	controllerAgentName = "in-memory-channel-controller"
+)
+
+// NewController initializes the controller and is called by the generated code.
+// Registers event handlers to enqueue events.
+func NewController(
+	opt reconciler.Options,
+	dispatcherNamespace string,
+	dispatcherDeploymentName string,
+	dispatcherServiceName string,
+	inmemorychannelinformer messaginginformers.InMemoryChannelInformer,
+	deploymentInformer appsv1informers.DeploymentInformer,
+	serviceInformer corev1informers.ServiceInformer,
+	endpointsInformer corev1informers.EndpointsInformer,
+) *controller.Impl {
+
+	r := &Reconciler{
+		Base:                     reconciler.NewBase(opt, controllerAgentName),
+		dispatcherNamespace:      dispatcherNamespace,
+		dispatcherDeploymentName: dispatcherDeploymentName,
+		dispatcherServiceName:    dispatcherServiceName,
+		inmemorychannelLister:    inmemorychannelinformer.Lister(),
+		inmemorychannelInformer:  inmemorychannelinformer.Informer(),
+		deploymentLister:         deploymentInformer.Lister(),
+		serviceLister:            serviceInformer.Lister(),
+		endpointsLister:          endpointsInformer.Lister(),
+	}
+	r.impl = controller.NewImpl(r, r.Logger, ReconcilerName)
+
+	r.Logger.Info("Setting up event handlers")
+	inmemorychannelinformer.Informer().AddEventHandler(controller.HandleAll(r.impl.Enqueue))
+
+	// Set up watches for dispatcher resources we care about, since any changes to these
+	// resources will affect our Channels. So, set up a watch here, that will cause
+	// a global Resync for all the channels to take stock of their health when these change.
+	deploymentInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
+		FilterFunc: controller.FilterWithNameAndNamespace(dispatcherNamespace, dispatcherDeploymentName),
+		Handler:    r,
+	})
+	serviceInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
+		FilterFunc: controller.FilterWithNameAndNamespace(dispatcherNamespace, dispatcherServiceName),
+		Handler:    r,
+	})
+	endpointsInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
+		FilterFunc: controller.FilterWithNameAndNamespace(dispatcherNamespace, dispatcherServiceName),
+		Handler:    r,
+	})
+	return r.impl
+}
