+++ pkg/reconciler/inmemorychannel/controller_test.go
@@ -1,0 +1,64 @@
+/*
+Copyright 2019 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package inmemorychannel
+
+import (
+	"testing"
+
+	fakeclientset "github.com/knative/eventing/pkg/client/clientset/versioned/fake"
+	informers "github.com/knative/eventing/pkg/client/informers/externalversions"
+	"github.com/knative/eventing/pkg/reconciler"
+	logtesting "github.com/knative/pkg/logging/testing"
+	kubeinformers "k8s.io/client-go/informers"
+	fakekubeclientset "k8s.io/client-go/kubernetes/fake"
+)
+
+func TestNewController(t *testing.T) {
+	kubeClient := fakekubeclientset.NewSimpleClientset()
+	eventingClient := fakeclientset.NewSimpleClientset()
+
+	// Create informer factories with fake clients. The second parameter sets the
+	// resync period to zero, disabling it.
+	kubeInformerFactory := kubeinformers.NewSharedInformerFactory(kubeClient, 0)
+	eventingInformerFactory := informers.NewSharedInformerFactory(eventingClient, 0)
+
+	// Eventing
+	imcInformer := eventingInformerFactory.Messaging().V1alpha1().InMemoryChannels()
+
+	// Kube
+	serviceInformer := kubeInformerFactory.Core().V1().Services()
+	endpointsInformer := kubeInformerFactory.Core().V1().Endpoints()
+	deploymentInformer := kubeInformerFactory.Apps().V1().Deployments()
+
+	c := NewController(
+		reconciler.Options{
+			KubeClientSet:     kubeClient,
+			EventingClientSet: eventingClient,
+			Logger:            logtesting.TestLogger(t),
+		},
+		systemNS,
+		dispatcherDeploymentName,
+		dispatcherServiceName,
+		imcInformer,
+		deploymentInformer,
+		serviceInformer,
+		endpointsInformer)
+
+	if c == nil {
+		t.Fatalf("Failed to create with NewController")
+	}
+}
